+++ src/provider/CodeActionProvider/otherCodeAction.ts
@@ -15,6 +15,7 @@ import { getFileAllFunc } from '../../tools/visitor/getFileAllFuncList';
 import type { showUnknownAnalyze } from '../CodeLens/showUnknownAnalyze';
 import { getFucDefWordUpFix } from '../Def/getFucDefWordUpFix';
 import { posAtFnRef } from '../Def/posAtFnRef';
+import { CodeActionAddGuiName } from './tools/CodeActionAddGuiName';
 import { CodeActionIncludePath } from './tools/CodeActionIncludePath';
 
 function atFnHead(
@@ -147,6 +148,8 @@ export function otherCodeAction(
     if (AhkTokenLine.detail.includes(EDetail.isDirectivesLine)) {
         need.push(...CodeActionIncludePath(active, AhkTokenLine, AhkFileData));
     }
+
+    need.push(...CodeActionAddGuiName(active, AhkTokenLine, AhkFileData));
 
     return need;
 }